refactor(lint): fix import ordering and unused variables - reduce Ruff errors from 98 to 75
- Fixed all F841 unused variable errors using Ruff's unsafe fixes
- Reorganized imports in key strategy and utility modules
- Added comprehensive Ruff linting to development workflow
- Enhanced cursor rules with Ruff standards and auto-commit prompts
- Updated dev scripts to focus on main directories (src, utils, scripts, tests)
- Improved docstrings and module organization across codebase

Remaining: 75 E402 import-ordering errors to be addressed in follow-up

+++ legacy_src/utils/data_utils.py
@@ -4,6 +4,7 @@ import pyarrow.parquet as pq
 import os
 from typing import List
 
+
 def csv_to_parquet(csv_path, parquet_path):
     """
     Convert a CSV file in the root-level data directory to Parquet format.
@@ -11,6 +12,7 @@ def csv_to_parquet(csv_path, parquet_path):
     df = pd.read_csv(csv_path)
     table = pa.Table.from_pandas(df)
     pq.write_table(table, parquet_path)
+
 
 def load_atm_call_options_from_csvs(csv_dir: str) -> pd.DataFrame:
     """
@@ -21,20 +23,32 @@ def load_atm_call_options_from_csvs(csv_dir: str) -> pd.DataFrame:
         pd.DataFrame: Cleaned DataFrame with only ATM CALL options for nearest expiry.
     """
     # Gather all CSV files
-    csv_files = [os.path.join(csv_dir, f) for f in os.listdir(csv_dir) if f.endswith('.csv')]
+    csv_files = [
+        os.path.join(csv_dir, f) for f in os.listdir(csv_dir) if f.endswith(".csv")
+    ]
     if not csv_files:
         raise FileNotFoundError(f"No CSV files found in {csv_dir}")
 
     # Load all CSVs into a single DataFrame
     df = pd.concat([pd.read_csv(f) for f in csv_files], ignore_index=True)
-    
+
     # Basic validation
     if df.empty:
         print("Warning: No data loaded from CSV files")
         return pd.DataFrame()
-    
+
     # Check required columns exist
-    required_columns = ['timestamp', 'expiry', 'strike', 'option_type', 'bid', 'ask', 'last', 'volume', 'open_interest']
+    required_columns = [
+        "timestamp",
+        "expiry",
+        "strike",
+        "option_type",
+        "bid",
+        "ask",
+        "last",
+        "volume",
+        "open_interest",
+    ]
     missing_columns = [col for col in required_columns if col not in df.columns]
     if missing_columns:
         print(f"Error: Missing required columns: {missing_columns}")
@@ -42,61 +56,62 @@ def load_atm_call_options_from_csvs(csv_dir: str) -> pd.DataFrame:
         return pd.DataFrame()
 
     # Ensure correct dtypes
-    df['timestamp'] = pd.to_datetime(df['timestamp'])
-    df['expiry'] = pd.to_datetime(df['expiry'])
-    df['strike'] = df['strike'].astype(float)
-    df['bid'] = df['bid'].astype(float)
-    df['ask'] = df['ask'].astype(float)
-    df['last'] = df['last'].astype(float)
-    df['volume'] = df['volume'].astype(float)
-    df['open_interest'] = df['open_interest'].astype(float)
+    df["timestamp"] = pd.to_datetime(df["timestamp"])
+    df["expiry"] = pd.to_datetime(df["expiry"])
+    df["strike"] = df["strike"].astype(float)
+    df["bid"] = df["bid"].astype(float)
+    df["ask"] = df["ask"].astype(float)
+    df["last"] = df["last"].astype(float)
+    df["volume"] = df["volume"].astype(float)
+    df["open_interest"] = df["open_interest"].astype(float)
 
     # Filter for nearest expiry
-    nearest_expiry = df['expiry'].min()
-    df = df[df['expiry'] == nearest_expiry]
+    nearest_expiry = df["expiry"].min()
+    df = df[df["expiry"] == nearest_expiry]
 
     # Only CALL options (handle both 'CALL' and 'CE' formats)
-    df = df[df['option_type'].str.upper().isin(['CALL', 'CE'])]
-    
+    df = df[df["option_type"].str.upper().isin(["CALL", "CE"])]
+
     if df.empty:
         print("Warning: No CALL options found after filtering")
         return pd.DataFrame()
 
     # For each timestamp, find ATM strike (strike closest to median strike)
     result_rows = []
-    
+
     # Get unique timestamps
-    unique_timestamps = df['timestamp'].unique()
-    
+    unique_timestamps = df["timestamp"].unique()
+
     for ts in unique_timestamps:
         # Get data for this timestamp
-        mask = df['timestamp'] == ts
+        mask = df["timestamp"] == ts
         group = df[mask].copy()  # Create a copy to avoid SettingWithCopyWarning
-        
+
         if len(group) == 0:
             continue
-            
+
         # Find the strike closest to the median strike (proxy for ATM)
-        median_strike = group['strike'].median()
-        group['strike_diff'] = abs(group['strike'] - median_strike)
-        atm_row = group.loc[group['strike_diff'].idxmin()]
-        
+        median_strike = group["strike"].median()
+        group["strike_diff"] = abs(group["strike"] - median_strike)
+        atm_row = group.loc[group["strike_diff"].idxmin()]
+
         # Convert to dictionary, excluding the temporary column
-        row_dict = atm_row.drop('strike_diff').to_dict()
+        row_dict = atm_row.drop("strike_diff").to_dict()
         result_rows.append(row_dict)
-    
+
     if not result_rows:
         print("Warning: No ATM rows found")
         return pd.DataFrame()
-    
+
     # Create DataFrame from list of dictionaries
     result_df = pd.DataFrame(result_rows)
-    
+
     # Sort by timestamp
-    result_df = result_df.sort_values('timestamp').reset_index(drop=True)
-    
+    result_df = result_df.sort_values("timestamp").reset_index(drop=True)
+
     return result_df
+
 
 # Example usage (uncomment to test)
 # df = load_atm_call_options_from_csvs('data/nse')
-# print(df.head()) 
+# print(df.head())
